added secrets.sh and docker-compose.yaml
The secrets.sh file sets up the passwords for the database and makes a copy of the gcp key. And the docker-compose.yaml has the configuration for the DB side of the app. However, there is placeholder data which will be updated once we get the main VM with gcp setup.